+++ 2017-18/Fürholzer_Kotlin_Presentation/Material/Exercises/Angabe.odt
