+++ 2017-18/Fürholzer_Kotlin_Presentation/Material/Exercises/Solutions/2D3DObjekte/JavaSource/main/Cube.java
@@ -1,0 +1,12 @@
+package main;
+
+public class Cube extends Prism{
+	private double len;
+	public Cube(double length) {
+		super(new Square(length),length);
+		len = length;
+	}
+	public String toString() {
+		return "Würfel [ Seitenlänge : "+this.len+" ]";
+	}
+}
