+++ 2017-18/Fürholzer_Kotlin_Presentation/Material/Exercises/Solutions/2D3DObjekte/JavaSource/main/Cuboid.java
@@ -1,0 +1,16 @@
+package main;
+
+public class Cuboid extends Prism{
+	private double len;
+	private double width;
+	private double height;
+	public Cuboid(double len,double width,double height) {
+		super(new Rectangle(len,width), height);
+		this.len=len;
+		this.height=height;
+		this.width = width;
+	}
+	public String toString() {
+		return "Rechteck [ Länge : "+len+" Höhe : "+height+ " Breite :"+width +" ]";
+	}
+}
