+++ 2017-18/Fürholzer_Kotlin_Presentation/Material/Exercises/Solutions/2D3DObjekte/JavaSource/main/Globe.java
@@ -1,0 +1,21 @@
+package main;
+
+public class Globe extends Geometric3dObject {
+	private double radius;
+	public Globe(double rad) {
+		this.radius=rad;
+	}
+	public double getNetArea() {
+		return Math.PI*this.radius*4;
+		//Keine Garantie auf Richtigkeit
+	}
+
+	public double getVolume() {
+		return Math.PI*this.radius*4/3;
+	}
+
+	public String toString() {
+		return "Kugel [ Radius : "+this.radius+" ]";
+	}
+
+}
