+++ 2017-18/Fürholzer_Kotlin_Presentation/Material/Exercises/Solutions/2D3DObjekte/KotlinSource/main/Circle.kt
