--- Fürholzer_Kotlin_Presentation/Material/Exercises/Solutions/2D3DObjekte/JavaSource/main/Cylinder.java
@@ -1,14 +1,0 @@
-package main;
-
-public class Cylinder extends Prism {
-	private double height;
-	private double rad;
-	public Cylinder(double rad,double height) {
-		super(new Circle(rad),height);
-		this.height=height;
-		this.rad=rad;
-	}
-	public String toString() {
-		return "Zylinder : [ Radius : "+this.rad+" Höhe : "+this.height+" ]";
-	}
-}
