--- Fürholzer_Kotlin_Presentation/Material/Exercises/Solutions/README.md
@@ -1,5 +1,0 @@
-# Important
-
-These solutions are created using Eclipse Oxygen Release (4.7.0) with Kotlin Plugin from the Eclipse Store.
-
-Furthermore I tried to comment important things about the language when possible.
